Add AI player using minimax algorithm, rendering game unwinnable
This commit introduces a new feature to the game: an AI player that uses the minimax algorithm to choose its moves. Unfortunately, this also means that the game is now impossible to win.

Some of the functions required to implement the AI player, such as checkWinner, were added to the GameController class. However, there's a question whether it would be better to move them to the GameBoard class instead. Further investigation and discussion are needed to determine the optimal implementation.

Changes made:
- Added a new AI player that uses the minimax algorithm
- Moved some functions required for the AI player to GameController
- Raised a question whether checkWinner should be moved to GameBoard

TODO:
- Investigate whether checkWinner should be moved to GameBoard
- Discuss and determine the optimal implementation with the team